@@ -1,74 +1,236 @@
 package net.tigereye.chestcavity;
 
-
+// import com.github.alexthe666.alexsmobs.AlexsMobs;
+import icyllis.modernui.mc.neoforge.MenuScreenFactory;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
-import net.minecraft.client.gui.ScreenManager;
-import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLDedicatedServerSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.SpawnPlacementTypes;
+import net.minecraft.world.level.levelgen.Heightmap;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.neoforged.fml.event.lifecycle.FMLDedicatedServerSetupEvent;
+import net.neoforged.fml.loading.FMLEnvironment;
+import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
+import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.AddReloadListenerEvent;
+import net.neoforged.neoforge.event.entity.RegisterSpawnPlacementsEvent;
+import net.tigereye.chestcavity.chestcavities.organs.OrganManager;
+import net.tigereye.chestcavity.chestcavities.types.json.GeneratedChestCavityAssignmentManager;
+import net.tigereye.chestcavity.chestcavities.types.json.GeneratedChestCavityTypeManager;
+import net.tigereye.chestcavity.client.command.ModernUIClientCommands;
+import net.tigereye.chestcavity.client.input.ModernUIKeyDispatcher;
+import net.tigereye.chestcavity.client.modernui.container.TestModernUIContainerFragment;
+import net.tigereye.chestcavity.client.modernui.container.TestModernUIContainerMenu;
+import net.tigereye.chestcavity.client.modernui.skill.SkillHotbarClientData;
+import net.tigereye.chestcavity.client.render.ChestCavityClientRenderers;
+import net.tigereye.chestcavity.command.ModernUIServerCommands;
+import net.tigereye.chestcavity.command.RecipeDebugCommands;
+import net.tigereye.chestcavity.compat.guzhenren.event.NoDropEvents;
+import net.tigereye.chestcavity.compat.guzhenren.util.hun_dao.soulbeast.command.SoulBeastCommands;
 import net.tigereye.chestcavity.config.CCConfig;
+import net.tigereye.chestcavity.engine.dot.DoTEngine;
+import net.tigereye.chestcavity.engine.reaction.ReactionEngine;
+import net.tigereye.chestcavity.guscript.command.GuScriptCommands;
+import net.tigereye.chestcavity.guscript.fx.client.FxClientHooks;
+import net.tigereye.chestcavity.guscript.fx.gecko.client.GeckoFxClient;
+import net.tigereye.chestcavity.guscript.registry.FxDefinitionLoader;
+import net.tigereye.chestcavity.guscript.registry.GeckoFxDefinitionLoader;
+import net.tigereye.chestcavity.guscript.registry.GuScriptFlowLoader;
+import net.tigereye.chestcavity.guscript.registry.GuScriptLeafLoader;
+import net.tigereye.chestcavity.guscript.registry.GuScriptRuleLoader;
+import net.tigereye.chestcavity.guscript.runtime.exec.GuScriptListenerHooks;
+import net.tigereye.chestcavity.guscript.runtime.flow.GuScriptFlowEvents;
+import net.tigereye.chestcavity.guscript.ui.GuScriptScreen;
+import net.tigereye.chestcavity.guzhenren.GuzhenrenModule;
+import net.tigereye.chestcavity.listeners.KeybindingClientListeners;
 import net.tigereye.chestcavity.network.NetworkHandler;
+import net.tigereye.chestcavity.network.ServerEvents;
 import net.tigereye.chestcavity.registration.*;
+import net.tigereye.chestcavity.skill.ActiveSkillRegistry;
+import net.tigereye.chestcavity.soul.command.SoulCommands;
+import net.tigereye.chestcavity.soul.entity.SoulEntityAttributes;
+import net.tigereye.chestcavity.soul.entity.SoulClanSpawner;
+import net.tigereye.chestcavity.soul.entity.TestSoulSpawner;
+import net.tigereye.chestcavity.soul.profile.capability.CapabilitySnapshots;
 import net.tigereye.chestcavity.ui.ChestCavityScreen;
+import net.tigereye.chestcavity.util.retention.OrganRetentionRules;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 @Mod(ChestCavity.MODID)
 public class ChestCavity {
-	public static final String MODID = "chestcavity";
-    private static final boolean DEBUG_MODE = false; //Make SURE that this is false when building
-	public static final Logger LOGGER = LogManager.getLogger();
-	public static CCConfig config;
-	public static final ResourceLocation COMPATIBILITY_TAG = new ResourceLocation(MODID,"organ_compatibility");
-	public static final ItemGroup ORGAN_ITEM_GROUP = new ItemGroup("chestcavity.organs") {
-		@Override
-		public ItemStack makeIcon() {
-			return new ItemStack(CCItems.HUMAN_STOMACH.get());
-		}
-	};
 
-	public ChestCavity() {
-		LOGGER.info("ChestCavity mod created, starting the loading sequence!");
-		printOnDebug("ChestCavity mod created!");
-		IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
-		bus.addListener(this::setup);
-		bus.addListener(this::doClientStuff);
+  public static final String MODID = "chestcavity";
+  private static final boolean DEBUG_MODE = false; // Make SURE that this is false when building
+  public static final Logger LOGGER = LogManager.getLogger();
+  public static CCConfig config;
+  public static final ResourceLocation COMPATIBILITY_TAG = id("organ_compatibility");
 
-		//Register mod resources
-		AutoConfig.register(CCConfig.class, GsonConfigSerializer::new);
-		config = AutoConfig.getConfigHolder(CCConfig.class).getConfig();
-		CCContainers.MENU_TYPES.register(bus);
-		CCItems.ITEMS.register(bus);
-		CCRecipes.RECIPE_SERIALIZERS.register(bus);
-		CCEnchantments.ENCHANTMENTS.register(bus);
-		CCListeners.register();
-		CCStatusEffects.MOB_EFFECTS.register(bus);
-		CCTagOrgans.init();
-}
+  public static ResourceLocation id(String path) {
+    return ResourceLocation.fromNamespaceAndPath(MODID, path);
+  }
 
-	public void setup(FMLCommonSetupEvent event) {
-		NetworkHandler.init();
-	}
+  public ChestCavity(IEventBus modEventBus) {
+    // AlexsMobs
+    IEventBus bus = modEventBus;
+    bus.addListener(this::setup);
+    bus.addListener(this::doClientStuff);
+    bus.addListener(this::registerMenuScreens);
+    bus.addListener(this::doServerStuff);
+    bus.addListener(NetworkHandler::registerCommon);
 
-	public void doClientStuff(FMLClientSetupEvent event) {
-		ScreenManager.register(CCContainers.CHEST_CAVITY_SCREEN_HANDLER.get(), ChestCavityScreen::new);
-		CCKeybindings.init();
-	}
+    NeoForge.EVENT_BUS.addListener(ServerEvents::onPlayerLogin);
+    NeoForge.EVENT_BUS.addListener(ServerEvents::onPlayerRespawn);
+    NeoForge.EVENT_BUS.addListener(ServerEvents::onPlayerClone);
+    NeoForge.EVENT_BUS.addListener(ServerEvents::onPlayerChangedDimension);
+    NeoForge.EVENT_BUS.addListener(ServerEvents::onLivingDeath);
+    NeoForge.EVENT_BUS.addListener(this::registerReloadListeners);
+    NeoForge.EVENT_BUS.addListener(GuScriptListenerHooks::onLivingDamage);
+    NeoForge.EVENT_BUS.addListener(GuScriptListenerHooks::onPlayerTick);
+    NeoForge.EVENT_BUS.addListener(GuScriptFlowEvents::onServerTick);
+    NeoForge.EVENT_BUS.addListener(GuScriptFlowEvents::onPlayerLogout);
+    NeoForge.EVENT_BUS.addListener(GuScriptCommands::register);
+    NeoForge.EVENT_BUS.addListener(RecipeDebugCommands::register);
+    NeoForge.EVENT_BUS.addListener(ModernUIServerCommands::register);
+    NeoForge.EVENT_BUS.addListener(SoulBeastCommands::register);
+    NeoForge.EVENT_BUS.addListener(SoulCommands::register);
+    // Central DoT manager ticking
+    DoTEngine.bootstrap();
+    NeoForge.EVENT_BUS.addListener(TestSoulSpawner::onServerTick);
+    SoulClanSpawner.init();
+    if (FMLEnvironment.dist.isClient()) {
+      NeoForge.EVENT_BUS.addListener(KeybindingClientListeners::onClientTick);
+      NeoForge.EVENT_BUS.addListener(ModernUIKeyDispatcher::onClientTick);
+      NeoForge.EVENT_BUS.addListener(GeckoFxClient::onClientTick);
+      NeoForge.EVENT_BUS.addListener(GeckoFxClient::onRenderLevel);
+      NeoForge.EVENT_BUS.addListener(ModernUIClientCommands::register);
+    }
 
-	public static boolean isDebugMode() {
-		return DEBUG_MODE;
-	}
+    if (FMLEnvironment.dist.isClient()) {
+      bus.addListener(this::registerClientReloadListeners);
+      bus.addListener(ChestCavityClientRenderers::onRegisterRenderers);
+    }
 
-	public static void printOnDebug(String stringToPrint) {
-		if(DEBUG_MODE) {
-			System.out.println("DEBUG: " + stringToPrint);
-		}
-	}
+    AutoConfig.register(CCConfig.class, GsonConfigSerializer::new);
+    config = AutoConfig.getConfigHolder(CCConfig.class).getConfig();
+    // Register mod resources
+    // AutoConfig.register(CCConfig.class, GsonConfigSerializer::new);
+    // config = AutoConfig.getConfigHolder(CCConfig.class).getConfig();
+    CCCreativeTabs.TABS.register(bus);
+    CCAttachments.ATTACHMENT_TYPES.register(bus);
+    CCContainers.MENU_TYPES.register(bus);
+    CCItems.ITEMS.register(bus);
+    CCEntities.ENTITY_TYPES.register(bus);
+    CCSoundEvents.SOUND_EVENTS.register(bus);
+    CCRecipes.RECIPE_SERIALIZERS.register(bus);
+    CCRecipes.RECIPE_TYPES.register(bus);
+    CCEnchantments.ENCHANTMENTS.register(bus);
+    CCListeners.register();
+    CCStatusEffects.MOB_EFFECTS.register(bus);
+    bus.addListener(CCKeybindings::register);
+    bus.addListener(SoulEntityAttributes::onAttributeCreation);
+    bus.addListener(this::registerSpawnPlacements);
+    CCTagOrgans.init();
+    CapabilitySnapshots.bootstrap();
+    net.tigereye.chestcavity.soul.runtime.SoulRuntimeHandlers.bootstrap();
+    OrganRetentionRules.registerNamespace(MODID);
+    if (ModList.get().isLoaded("guzhenren")) {
+      ActiveSkillRegistry.bootstrap();
+      GuzhenrenModule.bootstrap(bus, NeoForge.EVENT_BUS);
+      ReactionEngine.bootstrap();
+      // 提前注册古真人召唤相关的无掉落事件，避免召唤物产生战利品。
+      NoDropEvents.init();
+    }
+
+    // CCCommands.register();
+    // CCNetworkingPackets.register();
+    // ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new
+    // OrganManager());
+    // ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new
+    // GeneratedChestCavityTypeManager());
+    // ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new
+    // GeneratedChestCavityAssignmentManager());
+    // CrossModContent.register();
+
+  }
+
+  public void setup(FMLCommonSetupEvent event) {}
+
+  public void doClientStuff(FMLClientSetupEvent event) {
+    FxClientHooks.init();
+    event.enqueueWork(SkillHotbarClientData::initialize);
+  }
+
+  private void registerMenuScreens(RegisterMenuScreensEvent event) {
+    event.register(CCContainers.CHEST_CAVITY_SCREEN_HANDLER.get(), ChestCavityScreen::new);
+    event.register(CCContainers.GUSCRIPT_MENU.get(), GuScriptScreen::new);
+    event.register(
+        CCContainers.TEST_MODERN_UI_MENU.get(),
+        MenuScreenFactory.create(
+            menu -> new TestModernUIContainerFragment((TestModernUIContainerMenu) menu)));
+  }
+
+  public void doServerStuff(FMLDedicatedServerSetupEvent event) {
+    // ServerPlayNetworking.registerGlobalReceiver(CCNetworkingPackets.RECEIVED_UPDATE_PACKET_ID,
+    // (server, player, handler, buf, sender) -> {
+    //    Optional<ChestCavityEntity> optional = ChestCavityEntity.of(player);
+    //    optional.ifPresent(chestCavityEntity ->
+    // NetworkUtil.ReadChestCavityReceivedUpdatePacket(chestCavityEntity.getChestCavityInstance()));
+    // });
+    // ServerPlayNetworking.registerGlobalReceiver(CCNetworkingPackets.HOTKEY_PACKET_ID, (server,
+    // player, handler, buf, sender) -> {
+    //    Optional<ChestCavityEntity> optional = ChestCavityEntity.of(player);
+    //    optional.ifPresent(chestCavityEntity ->
+    // NetworkUtil.ReadChestCavityHotkeyPacket(chestCavityEntity.getChestCavityInstance(),buf));
+    // });
+  }
+
+  private void registerReloadListeners(AddReloadListenerEvent event) {
+
+    event.addListener(new OrganManager());
+    event.addListener(new GeneratedChestCavityTypeManager());
+    event.addListener(new GeneratedChestCavityAssignmentManager());
+    event.addListener(new GuScriptLeafLoader());
+    event.addListener(new GuScriptRuleLoader());
+    event.addListener(new GuScriptFlowLoader());
+    event.addListener(new net.tigereye.chestcavity.compat.guzhenren.gufang.GuFangRecipeLoader());
+    // FX definitions are client-only; do not register on server reload
+  }
+
+  private void registerClientReloadListeners(RegisterClientReloadListenersEvent event) {
+    event.registerReloadListener(new GuScriptLeafLoader());
+    event.registerReloadListener(new GuScriptRuleLoader());
+    event.registerReloadListener(new GeckoFxDefinitionLoader());
+    event.registerReloadListener(new FxDefinitionLoader());
+  }
+
+  public static boolean isDebugMode() {
+    return DEBUG_MODE;
+  }
+
+  public static void printOnDebug(String stringToPrint) {
+    if (DEBUG_MODE) {
+      System.out.println("DEBUG: " + stringToPrint);
+    }
+  }
+
+  public static void printOnDebug(java.util.function.Supplier<String> supplier) {
+    if (DEBUG_MODE) {
+      System.out.println("DEBUG: " + supplier.get());
+    }
+  }
+
+  private void registerSpawnPlacements(RegisterSpawnPlacementsEvent event) {
+    event.register(
+        CCEntities.TEST_SOUL.get(),
+        SpawnPlacementTypes.ON_GROUND,
+        Heightmap.Types.MOTION_BLOCKING_NO_LEAVES,
+        net.tigereye.chestcavity.soul.entity.TestSoulEntity::checkSpawnRules,
+        RegisterSpawnPlacementsEvent.Operation.REPLACE);
+  }
 }
