@@ -1,12 +1,10 @@
 package net.tigereye.chestcavity.chestcavities.organs;
 
-
-import net.minecraft.util.ResourceLocation;
-
 import java.util.HashMap;
 import java.util.Map;
+import net.minecraft.resources.ResourceLocation;
 
 public class OrganData {
-    public boolean pseudoOrgan;
-    public Map<ResourceLocation,Float> organScores = new HashMap<>();
+  public boolean pseudoOrgan;
+  public Map<ResourceLocation, Float> organScores = new HashMap<>();
 }
