@@ -3,7 +3,7 @@ package net.tigereye.chestcavity.chestcavities.organs;
 import com.google.gson.JsonArray;
 
 public class OrganJsonFormat {
-    String itemID;
-    boolean pseudoOrgan;
-    JsonArray organScores;
+  String itemID;
+  boolean pseudoOrgan;
+  JsonArray organScores;
 }
