@@ -3,6 +3,6 @@ package net.tigereye.chestcavity.chestcavities.types.json;
 import com.google.gson.JsonArray;
 
 public class ChestCavityAssignmentJsonFormat {
-    String chestcavity;
-    JsonArray entities;
+  String chestcavity;
+  JsonArray entities;
 }
