@@ -1,34 +1,42 @@
 package net.tigereye.chestcavity.chestcavities.types.json;
 
 import com.google.gson.Gson;
-import net.minecraft.resources.IResourceManager;
-import net.minecraft.resources.IResourceManagerReloadListener;
-import net.minecraft.util.ResourceLocation;
-import net.tigereye.chestcavity.ChestCavity;
-
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.util.HashMap;
 import java.util.Map;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.ResourceManager;
+import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
+import net.tigereye.chestcavity.ChestCavity;
 
-public class GeneratedChestCavityAssignmentManager implements IResourceManagerReloadListener {
-    private static final ResourceLocation RESOURCE_LOCATION = new ResourceLocation(ChestCavity.MODID, "entity_assignment");
-    private final ChestCavityAssignmentSerializer SERIALIZER = new ChestCavityAssignmentSerializer();
-    public static Map<ResourceLocation, ResourceLocation> GeneratedChestCavityAssignments = new HashMap<>();
+public class GeneratedChestCavityAssignmentManager implements ResourceManagerReloadListener {
+  private static final ResourceLocation RESOURCE_LOCATION = ChestCavity.id("entity_assignment");
+  private final ChestCavityAssignmentSerializer SERIALIZER = new ChestCavityAssignmentSerializer();
+  public static Map<ResourceLocation, ResourceLocation> GeneratedChestCavityAssignments =
+      new HashMap<>();
 
-    @Override
-    public void onResourceManagerReload(IResourceManager manager) {
-        GeneratedChestCavityAssignments.clear();
-        ChestCavity.LOGGER.info("Loading chest cavity assignments.");
-        for(ResourceLocation id : manager.listResources(RESOURCE_LOCATION.getPath(), path -> path.endsWith(".json"))) {
-            try(InputStream stream = manager.getResource(id).getInputStream()) {
+  @Override
+  public void onResourceManagerReload(ResourceManager manager) {
+    GeneratedChestCavityAssignments.clear();
+    ChestCavity.LOGGER.info("Loading chest cavity assignments.");
+    manager
+        .listResources(
+            RESOURCE_LOCATION.getPath(), location -> location.getPath().endsWith(".json"))
+        .forEach(
+            (id, resource) -> {
+              try (InputStream stream = resource.open()) {
                 Reader reader = new InputStreamReader(stream);
-                GeneratedChestCavityAssignments.putAll(SERIALIZER.read(id,new Gson().fromJson(reader,ChestCavityAssignmentJsonFormat.class)));
-            } catch(Exception e) {
-                ChestCavity.LOGGER.error("Error occurred while loading resource json " + id.toString(), e);
-            }
-        }
-        ChestCavity.LOGGER.info("Loaded "+GeneratedChestCavityAssignments.size()+" chest cavity assignments.");
-    }
+                GeneratedChestCavityAssignments.putAll(
+                    SERIALIZER.read(
+                        id, new Gson().fromJson(reader, ChestCavityAssignmentJsonFormat.class)));
+              } catch (Exception e) {
+                ChestCavity.LOGGER.error(
+                    "Error occurred while loading resource json " + id.toString(), e);
+              }
+            });
+    ChestCavity.LOGGER.info(
+        "Loaded " + GeneratedChestCavityAssignments.size() + " chest cavity assignments.");
+  }
 }
