@@ -1,10 +1,9 @@
 package net.tigereye.chestcavity.config;
 
+public class CCModMenuIntegration { // implements ModMenuApi {
 
-public class CCModMenuIntegration { //implements ModMenuApi {
-
-    //@Override
-    //public ConfigScreenFactory<?> getModConfigScreenFactory() {
-    //    return parent -> AutoConfig.getConfigScreen(CCConfig.class, parent).get();
-    //}
+  // @Override
+  // public ConfigScreenFactory<?> getModConfigScreenFactory() {
+  //    return parent -> AutoConfig.getConfigScreen(CCConfig.class, parent).get();
+  // }
 }
