@@ -1,10 +1,9 @@
 package net.tigereye.chestcavity.interfaces;
 
-
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 
 public interface CCHungerManagerInterface {
-    
-    public void ccEat(Item item, PlayerEntity player);
+
+  public void ccEat(Item item, Player player);
 }
