@@ -1,6 +1,7 @@
 package net.tigereye.chestcavity.interfaces;
 
 public interface CCStatusEffect {
-    boolean CC_IsHarmful();
-    boolean CC_IsBeneficial();
+  boolean CC_IsHarmful();
+
+  boolean CC_IsBeneficial();
 }
