@@ -1,6 +1,7 @@
 package net.tigereye.chestcavity.interfaces;
 
 public interface CCStatusEffectInstance {
-    int getDuration();
-    void CC_setDuration(int duration);
+  int getDuration();
+
+  void CC_setDuration(int duration);
 }
