@@ -1,11 +1,11 @@
 package net.tigereye.chestcavity.listeners;
 
 public class EffectiveFoodScores {
-    public float digestion = 0;
-    public float nutrition = 0;
+  public float digestion = 0;
+  public float nutrition = 0;
 
-    public EffectiveFoodScores(float digestion, float nutrition){
-        this.digestion = digestion;
-        this.nutrition = nutrition;
-    }
+  public EffectiveFoodScores(float digestion, float nutrition) {
+    this.digestion = digestion;
+    this.nutrition = nutrition;
+  }
 }
