@@ -1,9 +1,8 @@
 package net.tigereye.chestcavity.listeners;
 
-
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.potion.Effects;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.LivingEntity;
 import net.tigereye.chestcavity.ChestCavity;
 import net.tigereye.chestcavity.chestcavities.instance.ChestCavityInstance;
 import net.tigereye.chestcavity.interfaces.CCStatusEffect;
@@ -12,59 +11,78 @@ import net.tigereye.chestcavity.registration.CCOrganScores;
 
 public class OrganAddStatusEffectListeners {
 
-    public static EffectInstance callMethods(LivingEntity livingEntity, ChestCavityInstance chestCavityInstance, EffectInstance mobEffectInstance){
-        ApplyBuffPurging(livingEntity, chestCavityInstance, mobEffectInstance);
-        ApplyDetoxification(livingEntity, chestCavityInstance, mobEffectInstance);
-        ApplyWithered(livingEntity, chestCavityInstance, mobEffectInstance);
-        return mobEffectInstance;
-    }
+  public static MobEffectInstance callMethods(
+      LivingEntity livingEntity,
+      ChestCavityInstance chestCavityInstance,
+      MobEffectInstance mobEffectInstance) {
+    ApplyBuffPurging(livingEntity, chestCavityInstance, mobEffectInstance);
+    ApplyDetoxification(livingEntity, chestCavityInstance, mobEffectInstance);
+    ApplyFiltration(livingEntity, chestCavityInstance, mobEffectInstance);
+    ApplyWithered(livingEntity, chestCavityInstance, mobEffectInstance);
+    return mobEffectInstance;
+  }
 
-    private static EffectInstance ApplyBuffPurging(LivingEntity entity, ChestCavityInstance cc, EffectInstance instance) {
-        if(cc.getOrganScore(CCOrganScores.BUFF_PURGING) > 0
-                && ((CCStatusEffect)(instance.getEffect())).CC_IsBeneficial())
-        {
-            CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
-            ccInstance.CC_setDuration((int)(instance.getDuration()/
-                    (1+(ChestCavity.config.BUFF_PURGING_DURATION_FACTOR*cc.getOrganScore(CCOrganScores.BUFF_PURGING)))));
-        }
-        return instance;
+  private static MobEffectInstance ApplyBuffPurging(
+      LivingEntity entity, ChestCavityInstance cc, MobEffectInstance instance) {
+    if (cc.getOrganScore(CCOrganScores.BUFF_PURGING) > 0
+        && ((CCStatusEffect) (instance.getEffect())).CC_IsBeneficial()) {
+      CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
+      ccInstance.CC_setDuration(
+          (int)
+              (instance.getDuration()
+                  / (1
+                      + (ChestCavity.config.BUFF_PURGING_DURATION_FACTOR
+                          * cc.getOrganScore(CCOrganScores.BUFF_PURGING)))));
     }
+    return instance;
+  }
 
-    private static EffectInstance ApplyDetoxification(LivingEntity entity, ChestCavityInstance cc, EffectInstance instance) {
-        if(cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION) <= 0
-        || cc.getOrganScore(CCOrganScores.DETOXIFICATION) == cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION))
-        {
-            return instance;
-        }
-        CCStatusEffect ccStatusEffect = (CCStatusEffect)instance.getEffect();
-        if(ccStatusEffect.CC_IsHarmful()){
-            CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
-            float detoxRatio = cc.getOrganScore(CCOrganScores.DETOXIFICATION)/cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION);
-            ccInstance.CC_setDuration((int) Math.max(1,instance.getDuration() * 2 / (1 + detoxRatio)));
-        }
-        return instance;
+  private static MobEffectInstance ApplyDetoxification(
+      LivingEntity entity, ChestCavityInstance cc, MobEffectInstance instance) {
+    if (cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION) <= 0
+        || cc.getOrganScore(CCOrganScores.DETOXIFICATION)
+            == cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION)) {
+      return instance;
     }
+    CCStatusEffect ccStatusEffect = (CCStatusEffect) instance.getEffect();
+    if (ccStatusEffect.CC_IsHarmful()) {
+      CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
+      float detoxRatio =
+          cc.getOrganScore(CCOrganScores.DETOXIFICATION)
+              / cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.DETOXIFICATION);
+      ccInstance.CC_setDuration((int) Math.max(1, instance.getDuration() * 2 / (1 + detoxRatio)));
+    }
+    return instance;
+  }
 
-    private static EffectInstance ApplyFiltration(LivingEntity entity, ChestCavityInstance cc, EffectInstance instance) {
-        float filtrationDiff = cc.getOrganScore(CCOrganScores.FILTRATION) - cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.FILTRATION);
-        if(filtrationDiff > 0
-                && instance.getEffect() == Effects.POISON)
-        {
-            CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
-            ccInstance.CC_setDuration((int)(instance.getDuration()/
-                    (1+(ChestCavity.config.FILTRATION_DURATION_FACTOR*cc.getOrganScore(CCOrganScores.FILTRATION)))));
-        }
-        return instance;
+  private static MobEffectInstance ApplyFiltration(
+      LivingEntity entity, ChestCavityInstance cc, MobEffectInstance instance) {
+    float filtrationDiff =
+        cc.getOrganScore(CCOrganScores.FILTRATION)
+            - cc.getChestCavityType().getDefaultOrganScore(CCOrganScores.FILTRATION);
+    if (filtrationDiff > 0 && instance.getEffect() == MobEffects.POISON) {
+      CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
+      ccInstance.CC_setDuration(
+          (int)
+              (instance.getDuration()
+                  / (1
+                      + (ChestCavity.config.FILTRATION_DURATION_FACTOR
+                          * cc.getOrganScore(CCOrganScores.FILTRATION)))));
     }
+    return instance;
+  }
 
-    private static EffectInstance ApplyWithered(LivingEntity entity, ChestCavityInstance cc, EffectInstance instance) {
-        if(cc.getOrganScore(CCOrganScores.WITHERED) > 0
-                && instance.getEffect() == Effects.WITHER)
-        {
-            CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
-            ccInstance.CC_setDuration((int)(instance.getDuration()/
-                    (1+(ChestCavity.config.WITHERED_DURATION_FACTOR*cc.getOrganScore(CCOrganScores.WITHERED)))));
-        }
-        return instance;
+  private static MobEffectInstance ApplyWithered(
+      LivingEntity entity, ChestCavityInstance cc, MobEffectInstance instance) {
+    if (cc.getOrganScore(CCOrganScores.WITHERED) > 0 && instance.getEffect() == MobEffects.WITHER) {
+      CCStatusEffectInstance ccInstance = (CCStatusEffectInstance) instance;
+      ccInstance.CC_setDuration(
+          (int)
+              (instance.getDuration()
+                  / (1
+                      + (ChestCavity.config.WITHERED_DURATION_FACTOR
+                          * cc.getOrganScore(CCOrganScores.WITHERED)))));
     }
+    return instance;
+  }
 }
