@@ -1,18 +1,20 @@
 package net.tigereye.chestcavity.listeners;
 
 import com.mojang.datafixers.util.Pair;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.world.World;
+import java.util.List;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 import net.tigereye.chestcavity.chestcavities.instance.ChestCavityInstance;
 
-import java.util.List;
-
-public class OrganFoodEffectCallback {
-    //COMMENTED BY BOONELDAN
-    public static void addFoodEffects(List<Pair<EffectInstance, Float>> list, ItemStack itemStack, World world, LivingEntity entity, ChestCavityInstance cc) {
-        //OrganFoodEffectListeners.callMethods(list, itemStack, world, entity, cc);
-    }
+public class OrganFoodEffectCallback { // todo: event so other mods can add??
+  public static void addFoodEffects(
+      List<Pair<MobEffectInstance, Float>> list,
+      ItemStack itemStack,
+      Level world,
+      LivingEntity entity,
+      ChestCavityInstance cc) {
+    OrganFoodEffectListeners.callMethods(list, itemStack, world, entity, cc);
+  }
 }
-
