@@ -1,11 +1,10 @@
 package net.tigereye.chestcavity.listeners;
 
-
-import net.minecraft.entity.LivingEntity;
+import net.minecraft.world.entity.LivingEntity;
 import net.tigereye.chestcavity.chestcavities.instance.ChestCavityInstance;
 
 public class OrganUpdateCallback {
-    public static void organUpdate(LivingEntity entity, ChestCavityInstance cc) {
-        OrganUpdateListeners.callMethods(entity, cc);
-    }
+  public static void organUpdate(LivingEntity entity, ChestCavityInstance cc) {
+    OrganUpdateListeners.callMethods(entity, cc);
+  }
 }
