@@ -1,25 +1,27 @@
 package net.tigereye.chestcavity.mob_effect;
 
-
-
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.potion.EffectType;
+import net.minecraft.world.effect.MobEffectCategory;
+import net.minecraft.world.entity.LivingEntity;
 import net.tigereye.chestcavity.ChestCavity;
-import net.tigereye.chestcavity.registration.CCDamageSource;
+import net.tigereye.chestcavity.registration.CCDamageSources;
 
 public class OrganRejection extends CCStatusEffect {
 
-    public OrganRejection(){
-        super(EffectType.NEUTRAL, 0xC8FF00);
-    }
+  public OrganRejection() {
+    super(MobEffectCategory.NEUTRAL, 0xC8FF00);
+  }
 
-    public boolean isDurationEffectTick(int duration, int amplifier) {
-        return duration <= 1;
-    }
+  @Override
+  public boolean shouldApplyEffectTickThisTick(int duration, int amplifier) {
+    return duration <= 1;
+  }
 
-    public void applyEffectTick(LivingEntity entity, int amplifier) {
-        if(!(entity.level.isClientSide)){
-            entity.hurt(CCDamageSource.ORGAN_REJECTION, ChestCavity.config.ORGAN_REJECTION_DAMAGE);
-        }
+  @Override
+  public boolean applyEffectTick(LivingEntity entity, int amplifier) {
+    if (!entity.level().isClientSide()) {
+      entity.hurt(
+          CCDamageSources.organRejection(entity), ChestCavity.config.ORGAN_REJECTION_DAMAGE);
     }
+    return true;
+  }
 }
