@@ -1,30 +1,31 @@
 package net.tigereye.chestcavity.mob_effect;
 
-
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.ItemStack;
-import net.minecraft.potion.EffectType;
-import net.minecraft.util.FoodStats;
+import net.minecraft.world.effect.MobEffectCategory;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.food.FoodData;
 import net.tigereye.chestcavity.ChestCavity;
-import net.tigereye.chestcavity.registration.CCItems;
+import net.tigereye.chestcavity.registration.CCFoodComponents;
 
-public class Ruminating extends CCStatusEffect{
+public class Ruminating extends CCStatusEffect {
 
-    public Ruminating(){
-        super(EffectType.BENEFICIAL, 0xC8FF00);
+  public Ruminating() {
+    super(MobEffectCategory.BENEFICIAL, 0xC8FF00);
+  }
+
+  @Override
+  public boolean shouldApplyEffectTickThisTick(int duration, int amplifier) {
+    return duration % ChestCavity.config.RUMINATION_TIME == 1;
+  }
+
+  @Override
+  public boolean applyEffectTick(LivingEntity entity, int amplifier) {
+    if (entity instanceof Player player) {
+      if (!entity.level().isClientSide()) {
+        FoodData hungerManager = player.getFoodData();
+        hungerManager.eat(CCFoodComponents.CUD_FOOD_COMPONENT);
+      }
     }
-
-    public boolean isDurationEffectTick(int duration, int amplifier) {
-        return duration % ChestCavity.config.RUMINATION_TIME == 1;
-    }
-
-    public void applyEffectTick(LivingEntity entity, int amplifier) {
-        if(entity instanceof PlayerEntity){
-            if(!(entity.level.isClientSide)) {
-                FoodStats hungerManager = ((PlayerEntity) entity).getFoodData();
-                hungerManager.eat(CCItems.CUD.get(), new ItemStack(CCItems.CUD.get()));
-            }
-        }
-    }
+    return true;
+  }
 }
