@@ -1,87 +1,88 @@
 package net.tigereye.chestcavity.recipes;
 
-
-import net.minecraft.inventory.CraftingInventory;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.crafting.IRecipeSerializer;
-import net.minecraft.item.crafting.SpecialRecipe;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.world.World;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.CraftingBookCategory;
+import net.minecraft.world.item.crafting.CraftingInput;
+import net.minecraft.world.item.crafting.CustomRecipe;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.level.Level;
 import net.tigereye.chestcavity.registration.CCItems;
 import net.tigereye.chestcavity.registration.CCRecipes;
 import net.tigereye.chestcavity.util.CommonOrganUtil;
 
-public class InfuseVenomGland extends SpecialRecipe {
-    public InfuseVenomGland(ResourceLocation id) {
-        super(id);
-    }
+public class InfuseVenomGland extends CustomRecipe {
+  public InfuseVenomGland(CraftingBookCategory category) {
+    super(category);
+  }
 
-    @Override
-    public boolean matches(CraftingInventory craftingInventory, World world) {
-        boolean foundVenomGland = false;
-        boolean foundPotion = false;
-        for(int i = 0; i < craftingInventory.getWidth(); ++i) {
-            for(int j = 0; j < craftingInventory.getHeight(); ++j) {
-                ItemStack itemStack = craftingInventory.getItem(i + j * craftingInventory.getWidth());
-                if (itemStack.getItem() == CCItems.VENOM_GLAND.get()) {
-                    if(foundVenomGland){
-                        return false;
-                    }
-                    foundVenomGland = true;
-                }
-                else if(itemStack.getItem() == Items.POTION ||
-                        itemStack.getItem() == Items.SPLASH_POTION ||
-                        itemStack.getItem() == Items.LINGERING_POTION){
-                    if(foundPotion){
-                        return false;
-                    }
-                    foundPotion = true;
-                }
-            }
+  @Override
+  public boolean matches(CraftingInput input, Level world) {
+    boolean foundVenomGland = false;
+    boolean foundPotion = false;
+    for (ItemStack stack : input.items()) {
+      if (stack.isEmpty()) {
+        continue;
+      }
+      if (stack.is(CCItems.VENOM_GLAND.get())) {
+        if (foundVenomGland) {
+          return false;
         }
-        return foundVenomGland&&foundPotion;
+        foundVenomGland = true;
+      } else if (stack.is(Items.POTION)
+          || stack.is(Items.SPLASH_POTION)
+          || stack.is(Items.LINGERING_POTION)) {
+        if (foundPotion) {
+          return false;
+        }
+        foundPotion = true;
+      } else {
+        return false;
+      }
     }
+    return foundVenomGland && foundPotion;
+  }
 
-    @Override
-    public ItemStack assemble(CraftingInventory craftingInventory) {
-        ItemStack venomGland = null;
-        ItemStack potion = null;
-        ItemStack output = null;
-        for(int i = 0; i < craftingInventory.getWidth(); ++i) {
-            for(int j = 0; j < craftingInventory.getHeight(); ++j) {
-                ItemStack itemStack = craftingInventory.getItem(i + j * craftingInventory.getWidth());
-                if (itemStack.getItem() == CCItems.VENOM_GLAND.get()) {
-                    if(venomGland != null){
-                        return ItemStack.EMPTY;
-                    }
-                    venomGland = itemStack;
-                }
-                else if(itemStack.getItem() == Items.POTION ||
-                        itemStack.getItem() == Items.SPLASH_POTION ||
-                        itemStack.getItem() == Items.LINGERING_POTION){
-                    if(potion != null){
-                        return ItemStack.EMPTY;
-                    }
-                    potion = itemStack;
-                }
-            }
+  @Override
+  public ItemStack assemble(CraftingInput input, HolderLookup.Provider lookup) {
+    ItemStack venomGland = ItemStack.EMPTY;
+    ItemStack potion = ItemStack.EMPTY;
+    for (ItemStack stack : input.items()) {
+      if (stack.isEmpty()) {
+        continue;
+      }
+      if (stack.is(CCItems.VENOM_GLAND.get())) {
+        if (!venomGland.isEmpty()) {
+          return ItemStack.EMPTY;
         }
-        if(venomGland != null && potion != null){
-            output = venomGland.copy();
-            CommonOrganUtil.setStatusEffects(output, potion);
-            return output;
+        venomGland = stack.copy();
+      } else if (stack.is(Items.POTION)
+          || stack.is(Items.SPLASH_POTION)
+          || stack.is(Items.LINGERING_POTION)) {
+        if (!potion.isEmpty()) {
+          return ItemStack.EMPTY;
         }
+        potion = stack.copy();
+      } else {
         return ItemStack.EMPTY;
+      }
     }
+    if (!venomGland.isEmpty() && !potion.isEmpty()) {
+      ItemStack output = venomGland.copy();
+      CommonOrganUtil.setStatusEffects(output, potion);
+      return output;
+    }
+    return ItemStack.EMPTY;
+  }
 
-    @Override
-    public boolean canCraftInDimensions(int width, int height) {
-        return (width * height >= 2);
-    }
+  @Override
+  public boolean canCraftInDimensions(int width, int height) {
+    return width * height >= 2;
+  }
 
-    @Override
-    public IRecipeSerializer<?> getSerializer() {
-        return CCRecipes.INFUSE_VENOM_GLAND.get();
-    }
+  @Override
+  public RecipeSerializer<?> getSerializer() {
+    return CCRecipes.INFUSE_VENOM_GLAND.get();
+  }
 }
