@@ -1,120 +1,122 @@
 package net.tigereye.chestcavity.recipes;
 
-
-import net.minecraft.inventory.CraftingInventory;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.crafting.ICraftingRecipe;
-import net.minecraft.item.crafting.IRecipeSerializer;
-import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.util.NonNullList;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.world.World;
+import java.util.List;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.NonNullList;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.CraftingBookCategory;
+import net.minecraft.world.item.crafting.CraftingInput;
+import net.minecraft.world.item.crafting.CraftingRecipe;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraft.world.level.Level;
 import net.tigereye.chestcavity.registration.CCRecipes;
 
-public class SalvageRecipe implements ICraftingRecipe {
-    private final Ingredient input;
-    private int required;
-    private final ItemStack outputStack;
-    private final ResourceLocation id;
+public class SalvageRecipe implements CraftingRecipe {
+  private final CraftingBookCategory category;
+  private final Ingredient input;
+  private final int required;
+  private final ItemStack outputStack;
 
-    public SalvageRecipe(Ingredient input, int required, ItemStack outputStack, ResourceLocation id){
-        this.input = input;
-        this.required = required;
-        this.outputStack = outputStack;
-        this.id = id;
+  public SalvageRecipe(
+      CraftingBookCategory category, Ingredient input, int required, ItemStack outputStack) {
+    this.category = category;
+    this.input = input;
+    this.required = required;
+    this.outputStack = outputStack;
+  }
+
+  public Ingredient getInput() {
+    return input;
+  }
+
+  public int getRequired() {
+    return required;
+  }
+
+  public ItemStack getOutputPrototype() {
+    return outputStack.copy();
+  }
+
+  public Item getOutputItem() {
+    return outputStack.getItem();
+  }
+
+  public int getOutputCount() {
+    return outputStack.getCount();
+  }
+
+  @Override
+  public CraftingBookCategory category() {
+    return category;
+  }
+
+  @Override
+  public NonNullList<Ingredient> getIngredients() {
+    return NonNullList.withSize(required, input);
+  }
+
+  @Override
+  public boolean matches(CraftingInput input, Level level) {
+    int count = 0;
+    for (ItemStack stack : input.items()) {
+      if (stack.isEmpty()) {
+        continue;
+      }
+      if (this.input.test(stack)) {
+        count++;
+      } else {
+        return false;
+      }
     }
+    return count > 0 && count % required == 0;
+  }
 
-    public Ingredient getInput(){
-        return input;
+  @Override
+  public ItemStack assemble(CraftingInput input, HolderLookup.Provider lookup) {
+    int count = 0;
+    List<ItemStack> stacks = input.items();
+    for (ItemStack stack : stacks) {
+      if (stack.isEmpty()) {
+        continue;
+      }
+      if (this.input.test(stack)) {
+        count++;
+      } else {
+        return ItemStack.EMPTY;
+      }
     }
-
-    public int getRequired(){return required;}
-
-    @Override
-    public NonNullList<Ingredient> getIngredients() {
-        return NonNullList.withSize(required,input);
+    if (count == 0 || count % required != 0) {
+      return ItemStack.EMPTY;
     }
-
-    @Override
-    public boolean matches(CraftingInventory inv, World world) {
-        //ChestCavity.LOGGER.info("Attempting to match salvage recipe");
-        int count = 0;
-        ItemStack target;
-        for(int i = 0; i < inv.getContainerSize(); ++i) {
-            target = inv.getItem(i);
-            if(target != null && target != ItemStack.EMPTY && target.getItem() != Items.AIR) {
-                if (input.test(inv.getItem(i))) {
-                    count++;
-                    //ChestCavity.LOGGER.info("Salvage recipe counts "+count+" "+target.getName()+"s");
-                }
-                else{
-                    //ChestCavity.LOGGER.info("Salvage recipe found bad item: "+target.getName().getString());
-                    return false;
-                }
-            }
-        }
-        //if(count > 0){
-            //ChestCavity.LOGGER.info("Found salvage recipe match");
-        //}
-        return count > 0 && count % required == 0;
+    int resultCount = (count / required) * outputStack.getCount();
+    if (resultCount > outputStack.getMaxStackSize()) {
+      return ItemStack.EMPTY;
     }
+    ItemStack output = outputStack.copy();
+    output.setCount(resultCount);
+    return output;
+  }
 
-    @Override
-    public ItemStack assemble(CraftingInventory inv) {
-        //ChestCavity.LOGGER.info("Attempting to craft salvage recipe");
-        int count = 0;
-        ItemStack target;
-        for(int i = 0; i < inv.getContainerSize(); ++i) {
-            target = inv.getItem(i);
-            if(target != null && target != ItemStack.EMPTY && target.getItem() != Items.AIR) {
-                if (input.test(inv.getItem(i))) {
-                    count++;
-                }
-                else{
-                    return ItemStack.EMPTY;
-                }
-            }
-        }
-        //ChestCavity.LOGGER.info("Found salvage recipe, count "+count);
-        if(count == 0 || count % required != 0){
-            //ChestCavity.LOGGER.info("Salvage recipe failed modulo check");
-            return ItemStack.EMPTY;
-        }
-        count = (count / required) * outputStack.getCount();
-        if(count > outputStack.getMaxStackSize()) {
-            //ChestCavity.LOGGER.info("Salvage recipe exceeded max stack size");
-            return ItemStack.EMPTY;
-        }
-        ItemStack out = getResultItem();
-        out.setCount(count);
-        return out;
-    }
+  @Override
+  public boolean canCraftInDimensions(int width, int height) {
+    return width * height >= required;
+  }
 
-    @Override
-    public boolean canCraftInDimensions(int width, int height) {
-        return true;
-    }
+  @Override
+  public ItemStack getResultItem(HolderLookup.Provider lookup) {
+    return outputStack.copy();
+  }
 
-    @Override
-    public ItemStack getResultItem() {
-        return outputStack.copy();
-    }
+  @Override
+  public RecipeSerializer<?> getSerializer() {
+    return CCRecipes.SALVAGE_RECIPE_SERIALIZER.get();
+  }
 
-    @Override
-    public ResourceLocation getId() {
-        return id;
-    }
-
-    @Override
-    public IRecipeSerializer<?> getSerializer() {
-        return CCRecipes.SALVAGE_RECIPE_SERIALIZER.get();
-    }
-
-    /*@Override
-    public RecipeType<SalvageRecipe> getType() {
-        return CCRecipes.SALVAGE_RECIPE_TYPE;
-    }
-    */
-
+  @Override
+  public RecipeType<?> getType() {
+    return CCRecipes.SALVAGE_RECIPE_TYPE.get();
+  }
 }
