@@ -3,8 +3,9 @@ package net.tigereye.chestcavity.recipes.json;
 import com.google.gson.JsonObject;
 
 public class SalvageRecipeJsonFormat {
-    JsonObject ingredient;
-    int required;
-    String result;
-    int count;
+  JsonObject ingredient;
+  int required;
+  String result;
+  int count;
+  String category;
 }
