@@ -1,16 +1,15 @@
 package net.tigereye.chestcavity.registration;
 
-import net.tigereye.chestcavity.listeners.OrganActivationListeners;
+import net.tigereye.chestcavity.listeners.*;
 
 public class CCListeners {
-    public static void register(){
-        //LootRegister.register();
-        //OrganUpdateListeners.register();
-        //OrganTickListeners.register();
-        OrganActivationListeners.register();
-        //OrganAddStatusEffectListeners.register();
-        //OrganFoodListeners.register();
-        //OrganFoodEffectListeners.register();
-        //OrganOnHitListeners.register();
-    }
+  public static void register() {
+    // OrganUpdateListeners.register();
+    // OrganTickListeners.register();
+    OrganActivationListeners.register();
+    // OrganAddStatusEffectListeners.register();
+    // OrganFoodListeners.register();
+    // OrganFoodEffectListeners.register();
+    // OrganOnHitListeners.register();
+  }
 }
