@@ -1,52 +1,14 @@
 package net.tigereye.chestcavity.registration;
 
-
-import net.minecraft.util.ResourceLocation;
+import net.minecraft.resources.ResourceLocation;
 import net.tigereye.chestcavity.ChestCavity;
 
 public class CCNetworkingPackets {
-    public static final ResourceLocation ORGAN_DATA_PACKET_ID = new ResourceLocation(ChestCavity.MODID,"organ_data");
-    public static final ResourceLocation UPDATE_PACKET_ID = new ResourceLocation(ChestCavity.MODID,"update");
-    public static final ResourceLocation RECEIVED_UPDATE_PACKET_ID = new ResourceLocation(ChestCavity.MODID,"received_update");
+  public static final ResourceLocation ORGAN_DATA_PACKET_ID = ChestCavity.id("organ_data");
+  public static final ResourceLocation UPDATE_PACKET_ID = ChestCavity.id("update");
+  public static final ResourceLocation RECEIVED_UPDATE_PACKET_ID =
+      ChestCavity.id("received_update");
+  public static final ResourceLocation HOTKEY_PACKET_ID = ChestCavity.id("hotkey");
 
-    public static final ResourceLocation HOTKEY_PACKET_ID = new ResourceLocation(ChestCavity.MODID, "hotkey");
-
-    public static void register() {
-        //COMMENTED BY BOONELDAN
-        //ServerLoginConnectionEvents.QUERY_START.register(NetworkUtil::sendOrganDataPacket);
-        //ServerPlayNetworking.registerGlobalReceiver(CCNetworkingPackets.RECEIVED_UPDATE_PACKET_ID, (server, player, handler, buf, sender) -> {
-        //    Optional<ChestCavityEntity> optional = ChestCavityEntity.of(player);
-        //    optional.ifPresent(chestCavityEntity -> NetworkUtil.ReadChestCavityReceivedUpdatePacket(chestCavityEntity.getChestCavityInstance()));
-        //});
-
-
-
-
-
-
-        //DONE
-        //ServerPlayNetworking.registerGlobalReceiver(CCNetworkingPackets.HOTKEY_PACKET_ID, (server, player, handler, buf, sender) -> {
-        //    Optional<ChestCavityEntity> optional = ChestCavityEntity.of(player);
-        //    optional.ifPresent(chestCavityEntity -> NetworkUtil.ReadChestCavityHotkeyPacket(chestCavityEntity.getChestCavityInstance(),buf));
-        //});
-        //ServerLoginNetworking.registerGlobalReceiver(CCNetworkingPackets.ORGAN_DATA_PACKET_ID,(server, handler, understood, buf, synchronizer, sender) -> {
-        //});
-    }
-
-    public static void registerClient(){
-        //DONE
-        //ClientPlayNetworking.registerGlobalReceiver(CCNetworkingPackets.UPDATE_PACKET_ID, (client, handler, buf, responseSender) -> {
-        //    Optional<ChestCavityEntity> optional = ChestCavityEntity.of(client.cameraEntity);
-        //    optional.ifPresent(chestCavityEntity -> NetworkUtil.ReadChestCavityUpdatePacket(chestCavityEntity.getChestCavityInstance(), buf));
-        //});
-
-
-
-
-        //COMMENTED BY BOONELDAN
-        //ClientLoginNetworking.registerGlobalReceiver(CCNetworkingPackets.ORGAN_DATA_PACKET_ID, (client, handler, buf, responseSender) -> {
-        //    NetworkUtil.readOrganDataPacket(buf);
-        //    return CompletableFuture.completedFuture(PacketByteBufs.empty());
-        //});
-    }
+  public static void register() {}
 }
