@@ -1,28 +1,27 @@
 package net.tigereye.chestcavity.util;
 
 public class Pair<A, B> {
-    private A left;
-    private B right;
+  private A left;
+  private B right;
 
-    public Pair(A left, B right) {
-        this.left = left;
-        this.right = right;
-    }
+  public Pair(A left, B right) {
+    this.left = left;
+    this.right = right;
+  }
 
-    public A getLeft() {
-        return this.left;
-    }
+  public A getLeft() {
+    return this.left;
+  }
 
-    public void setLeft(A left) {
-        this.left = left;
-    }
+  public void setLeft(A left) {
+    this.left = left;
+  }
 
-    public B getRight() {
-        return this.right;
-    }
+  public B getRight() {
+    return this.right;
+  }
 
-    public void setRight(B right) {
-        this.right = right;
-    }
+  public void setRight(B right) {
+    this.right = right;
+  }
 }
-
